Moved Overflow component from banner into UI-Components
